+++ libinstpatch/IpatchSF2VoiceCache_Gig.c
@@ -235,7 +235,7 @@ _gig_inst_to_sf2_voice_cache_convert (IpatchConverter *converter, GError **err)
 	      voice->loop_start = sample_info->loop_start;
 	      voice->loop_end = sample_info->loop_end;
 	      voice->root_note = sample_info->root_note;
-	      voice->fine_tune = sample_info->fine_tune;
+	      voice->fine_tune = (guint8)sample_info->fine_tune;
 
 	      switch (sample_info->options & IPATCH_DLS2_SAMPLE_LOOP_MASK)
 		{
